feat(components): add prepare script
