+++ modules/components/package.json
@@ -6,12 +6,14 @@
     "start": "dumi dev",
     "build:dist": "webpack",
     "build": "gulp compile",
-    "watch-build": "gulp --gulpfile gulpfile-watch.js",
+    "build:watch": "gulp --gulpfile gulpfile-build.js --watch",
+    "build:noDeclare": "gulp --gulpfile gulpfile-build.js",
     "docs:build": "dumi build",
     "prettier": "prettier -c --write .",
     "precommit": "lint-staged",
     "lint": "eslint --ext .ts,.tsx,.js,.jsx src",
     "test": "jest",
+    "prepare": "npm run build:noDeclare",
     "test:coverage": "jest --coverage",
     "publish:patch": "npm run build && npm version patch && npm publish --registry https://registry.npmjs.org"
   },
